move IconImage to separate folder
- icons are now autogenerated on Styleguide page but this component is not regular icon which is causing issue in displaying it in the list of icons
