+++ project-base/storefront/components/Layout/Header/AutocompleteSearch/AutocompleteSearchPopup.tsx
@@ -5,7 +5,7 @@ import {
     AUTOCOMPLETE_ARTICLE_LIMIT,
 } from './constants';
 import { ExtendedNextLink } from 'components/Basic/ExtendedNextLink/ExtendedNextLink';
-import { IconImage } from 'components/Basic/Icon/IconImage';
+import { IconImage } from 'components/Basic/IconImage/IconImage';
 import { Image } from 'components/Basic/Image/Image';
 import { Button } from 'components/Forms/Button/Button';
 import { useDomainConfig } from 'components/providers/DomainConfigProvider';
